refactor: Separation of UI and business logic about HelpModal
