+++ client/src/components/help/HelpModalPageUI.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react';
 import { DotLottieReact } from '@lottiefiles/dotlottie-react';
+import Indicator from './Indicator';
 import { HelpPageProps } from '@/types/help.types';
 import { cn } from '@/utils/cn';
 
@@ -37,22 +38,7 @@ const HelpPage = ({ pageData, isModalOpened, dotLottieRefCallback, pagenation, s
           </div>
         </article>
       </section>
-      <div className="relative top-5 flex flex-row items-center justify-center p-5">
-        {pagenation.map((isSelect, i) => {
-          return (
-            <button
-              key={i}
-              className={cn(
-                'mx-1.5 box-content h-2.5 w-2.5 rounded-full md:h-3 md:w-3',
-                isSelect ? 'border-4 border-halfbaked-300 bg-chartreuseyellow-300' : 'bg-eastbay-600',
-              )}
-              onClick={() => {
-                setPageIndex(i);
-              }}
-            />
-          );
-        })}
-      </div>
+      <Indicator pageData={pageData} pagenation={pagenation} setPageIndex={setPageIndex} />
     </div>
   );
 };
